Rename component directories to leave off the onos- prefix
