+++ pkg/e2t/connections.go
